resources: Update GPUFS disk to ROCm 6.1, HCL format (#29)
Update the disk image creation scripts to use the HCL packer formatted,
modified from the x86-ubuntu gem5 resource. The major feature of this is
a simplified one-step command to build the disk image.

In addition the GPU package versions are bumped to ROCm 6.1, PyTorch
2.2.2, and TensorFlow 2.14.0

The previous ROCm 5.4.2 resource is removed and the new disk image uses
an unversioned x86-ubuntu-gpu-ml directory.