resources: Update pannotia Makefiles for GPUFS/HW (#45)
There is an issue with the compiler used with GPUFS and hardware that
causes a hang or crash in fstream. PR #44 changed fstream to fopen as
fstream has similar issues in ROCm 4.0. To compromise and support both
versions, this PR updates the GPUFS and hardware (Makefile.default)
Makefiles to build the input parser with optimizations off (-O0) which
"solves" the issue. Since we typically do not care about simulation time
of reading input files, this seems like a decent compromise.

The GPUFS Makefiles also change the GEM5_PATH environment variable to
GEM5_ROOT to be consistent with each Makefile.gem5-fusion.

The BC application is unchanged as it implements its open parser in its
header file which cannot be compiled to an object file without
refactoring the benchmark.